requirements_security.txt
# Security Packages for SuperParaguai E-commerce Platform
# Install with: pip install -r requirements_security.txt

# Rate Limiting
django-ratelimit==4.1.0

# Login Attempt Tracking & Brute Force Protection
django-axes==6.1.0

# Additional Security
django-csp==3.7

# Input Validation & Sanitization
bleach==6.1.0

# Security Headers
django-security==0.15.0

# CORS Security
django-cors-headers==4.3.1

# JWT Security
djangorestframework-simplejwt==5.3.0

# File Upload Security
python-magic==0.4.27

# Password Security
django-password-validators==0.3.1

# Rate Limiting (Alternative)
django-ratelimit2==0.1.0

# Security Monitoring
django-security-logger==0.1.0
